chore: Rename execute command to verify
